+++ docs/en/changes/changes.md
@@ -64,7 +64,8 @@
 * Fix query `getGlobalTopology` throw exception when didn't find any services by the given Layer.
 * Fix the previous analysis result missing in the ALS `k8s-mesh` analyzer.
 * Fix `findEndpoint` query require `keyword` when using BanyanDB.
-* Adapt BanyanDB Java Client 0.7.0.
+* Support to analysis the ztunnel mapped IP address in eBPF Access Log Receiver.
+* Adapt BanyanDB Java Client 0.7.0-rc3.
 
 #### UI